+++ allure.test/debug.ts
@@ -1,6 +1,6 @@
 // import './suite/create.allure';
 // import './suite/valid.request.path.allure';
-// import './suite/valid.request.body.oas3.allure';
+import './suite/valid.request.body.oas3.allure';
 // import './suite/valid.request.body.oas2.allure';
 // import './suite/valid.request.contentType.oas3.allure';
 // import './suite/valid.request.contentType.oas2.allure';
@@ -11,4 +11,4 @@
 // import './suite/valid.request.para.allure';
 // import './suite/routerPrefix.allure';
 // import './suite/x-mute-env.allure';
-import './suite/x-mute-env.global.allure';
+// import './suite/x-mute-env.global.allure';